+++ framework/i18n/src/main/java/is/codion/framework/i18n/FrameworkMessages.java
@@ -78,8 +78,6 @@ public final class FrameworkMessages {
 
 	private static final String SETTINGS = "settings";
 	private static final String SELECT_INPUT_FIELD = "select_input_field";
-	private static final String SELECT_SEARCH_FIELD = "select_search_field";
-	private static final String SELECT_FILTER_FIELD = "select_filter_field";
 
 	private FrameworkMessages() {}
 
@@ -397,20 +395,6 @@ public final class FrameworkMessages {
 	 */
 	public static String selectInputField() {
 		return get(SELECT_INPUT_FIELD);
-	}
-
-	/**
-	 * @return select search field
-	 */
-	public static String selectSearchField() {
-		return get(SELECT_SEARCH_FIELD);
-	}
-
-	/**
-	 * @return select filter field
-	 */
-	public static String selectFilterField() {
-		return get(SELECT_FILTER_FIELD);
 	}
 
 	private static String get(String key) {